EN: changed usages of answering commands KP-8034

+++ src/Core/SharedKernels/Enumerator/Enumerator/ViewModels/InterviewDetails/Questions/RealQuestionViewModel.cs
@@ -152,7 +152,7 @@ namespace WB.Core.SharedKernels.Enumerator.ViewModels.InterviewDetails.Questions
                 questionId: this.questionIdentity.Id,
                 rosterVector: this.questionIdentity.RosterVector,
                 answerTime: DateTime.UtcNow,
-                answer: this.Answer.Value);
+                answer: (double) this.Answer.Value);
 
             try
             {
